fix: length_histogram[] must be initialized
NOTE: length_histogram[] *must* be initialized to 0 since it has:
1) automatic storage duration,
2) compute_bit_length function may not set length_histogram[i] for all i
3)  codewords_in_tree += length_histogram[i]; may add an uninitialized
value to codewords_in_tree afterwards.

Without this initialization, cosim will "mysteriously" run indefinitely.  But
it's not really mysterious because in the RTL cosim the uninit value may
be a large value which is outside an end of cosim condition check.