feat: add config_csim_prefix - global on clang sanitizer
add config_csim_prefix key support to ini file.  This
allows one to add a config_csim command to ALL
components if placed in the [DEFAULTS] section.

Can also add to individual components as well.

This will update tcl script so that it prefixes standard
config_csim command with this command.